feat(digitraffic): fix tests, add exports, directory structure (#99)
* feat!(digitraffic): move station path to base directory (#94)

* feat\!(digitraffic): remove utils/get_station_path

This was not generic enough to be considered an utility. Move it to  `base` instead and export through other means

* feat: add two new exports

digitraffic error class — this is a somewhat stable api that has many use cases
get station path — moved the export from /utils to index as a side-effect of removing /utils directory

* test: remove utils directory

* refactor: move get_station_path to base

* test: move get_station_path test here

* refactor: use index as package entry point for

digitraffic error, get station path

* refactor(digitraffic): move data directory to mocks (#93)

* test(digitraffic): fix type errors (#92)

* test: document expected ts errors, refactor

* test: document expected ts errors, nullish coalescing, use before assign

* test: refactor to call expect after promise is resolved

* feat: move station path to base (#96)

* feat\!(digitraffic): remove utils/get_station_path

This was not generic enough to be considered an utility. Move it to  `base` instead and export through other means

* feat: add two new exports

digitraffic error class — this is a somewhat stable api that has many use cases
get station path — moved the export from /utils to index as a side-effect of removing /utils directory

* test: remove utils directory

* refactor: move get_station_path to base

* test: move get_station_path test here

* refactor: use index as package entry point for

digitraffic error, get station path

* refactor: fix typo in doc comment (#95)

* fix: wrong import (#98)