fix: timetable row animation was quicker than intended (#123)
* fix: replace exit before enter with mode = await

* fix: slow down navigation indicator

On navigation back to /station, the station an user last pressed is animated and scrolled to.

Recently a new fadein animation was added. As it is shorter in duration and tied to the aforementioned animation, the navigation indicator animation turned into a flash from a subtle indicator.

In this fix, the durations are separated from eachother.