ci: revert use of pnpm fetch introduced in #143
`pnpm install --offline` uses cache to find dependencies. Currently, some dependencies are not included in the cache (although the lockfile is up to date) and the CI fails because of `ERR_PNPM_NO_OFFLINE_META`.
This is seems to be related to issue https://togithub.com/pnpm/pnpm/issues/5315. What's peculiar is that building the image locally works, but in a devcontainer or CI fails.

This is a hotfix and not a complete solution.