ci: add bash (#169)
Alpine doesn't come with bash installed. This resulted in the build command failing with the error message  sh: ./copy.sh: not found. The shebang #!/bin/bash in copy.sh was the cause of this issue as bash was not installed.

Solution: install bash; we can't use sh as we are using bash's read which is different than the one provided by sh.