feat: show nearby only if geolocation triggered navigation
We use query parameter for this which could be just typed in and break the component (e.g. if the module variable is not updated), but for now it's not a concern. Another way to implement is to share the state somehow which could be good in a future refactor.