feat(digitraffic): add use cached train
Motivation: the way train page is tied together makes it depend on a train being defined.
`useLiveTrains` probably has the data we need so we can render the page instantly rather than
wait for a graphql request to return data. On a slow connection this can mean that the content
is rendered a second faster than before resulting in better UX.

This is not a holy grail solution however. If the cache is garbage collected this hook kinda just is.