feat(android): status bar
the implementation is a bit spotty still as toggling theme via toggle is not reflected on the status bar. Edge-to-edge https://developer.android.com/develop/ui/views/layout/edge-to-edge could be considered, but there's an issue with both iOS and Android not using the env(safe-area-inset-top) or I could not get it to work.